requirements.txt
Flask==2.3.2
flask_sqlalchemy==3.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy~=2.0.27
flask-jwt-extended
passlib~=1.7.4
itsdangerous~=2.1.2
flask-migrate
gunicorn==21.2.0
Flask_WTF==1.2.1
WTForms==3.1.2
alembic~=1.14.1
Flask_CKEditor==0.5.1
Bootstrap_Flask==2.3.3
email_validator==2.1.0.post1
plotly
Flask_Login==0.6.3
apscheduler
Flask-Mail
flask[async]
pdfcrowd
logging~=0.4.9.6
Flask-Login==0.6.3
pytz
flask_cors
flask_socketio
firebase-admin